updated franchaisor and frachaise functionality

+++ routes/api.php
@@ -3,6 +3,9 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\Country\CountryController;
+use App\Http\Controllers\Dashboard\DashboardController;
+use App\Http\Controllers\Franchaisee\FranchaiseeController;
 use App\Http\Controllers\Franchaisor\FranchaisorController;
 
 Route::get('/user', function (Request $request) {
@@ -11,13 +14,25 @@ Route::get('/user', function (Request $request) {
 
 Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
+Route::post('/franchaisee-request', [FranchaiseeController::class, 'store']);
+Route::post('/franchaisor-request', [FranchaisorController::class, 'store']);
+Route::get('/franchaisors', [FranchaiseeController::class, 'index']);
 
-Route::group(['middleware' => 'auth:sanctum'], function () {
+
+Route::group(['middleware' => 'auth:sanctum' , 'role:admin'], function () {
     Route::get('/user', [AuthController::class, 'getUser']);
     Route::post('/logout', [AuthController::class, 'logout']);
     Route::post('/update', [AuthController::class, 'updateUser']);
-    Route::post('/franchaisor', [FranchaisorController::class, 'store']);
-    Route::get('/franchaisor/{id}', [FranchaisorController::class, 'show']);
-    Route::patch('/franchaisor/{id}', [FranchaisorController::class, 'update']);
-    Route::delete('/franchaisor/{id}', [FranchaisorController::class, 'destroy']);
+    Route::get('/countries', [CountryController::class, 'index']);
+    Route::get('/admin/franchaisors', [FranchaisorController::class, 'index']);
+    Route::post('/admin/franchaisor', [FranchaisorController::class, 'store']);
+    Route::get('/admin/franchaisor/{id}', [FranchaisorController::class, 'show']);
+    Route::patch('/admin/franchaisor/{id}', [FranchaisorController::class, 'update']);
+    Route::delete('/admin/franchaisor/{id}', [FranchaisorController::class, 'destroy']);
+    Route::get('/admin/franchaisees', [FranchaiseeController::class, 'index']);
+    Route::post('/admin/franchaisee', [FranchaiseeController::class, 'store']);
+    Route::get('/admin/franchaisee/{id}', [FranchaiseeController::class, 'show']);
+    Route::patch('/admin/franchaisee/{id}', [FranchaiseeController::class, 'update']);
+    Route::delete('/admin/franchaisee/{id}', [FranchaiseeController::class, 'destroy']);
+    Route::get('/admin/dashboard', [DashboardController::class, 'index']);
 });
